% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutocorrelationFunction.R
\name{AutocorrelationFunction}
\alias{AutocorrelationFunction}
\title{AutocorrelationFunction}
\usage{
AutocorrelationFunction(stat.data = NULL, dependent.variable = NULL)
}
\arguments{
\item{stat.data}{the dataset name as string}

\item{dependent.variable}{the dependent variable column name as string}
}
\value{
numeric
}
\description{
Function to calculate Autocorrelation
}
\details{
Call via \code{AutocorrelationFunction(...)}
}
\author{
Revanth Nemani, \email{raynemani@gmail.com}
}
