% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{AutoGrade}
\alias{AutoGrade}
\title{AutoGrade}
\usage{
AutoGrade(x)
}
\arguments{
\item{x}{the auto correlation value}
}
\value{
numeric
}
\description{
Function to grade autocorrelation result
}
\details{
Call via \code{AutoGrade(x)}
}
\author{
Revanth Nemani, \email{raynemani@gmail.com}
}
