% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{AutoGradeDescription}
\alias{AutoGradeDescription}
\title{AutoGradeDescription}
\usage{
AutoGradeDescription(x)
}
\arguments{
\item{x}{the auto correlation value}
}
\value{
numeric
}
\description{
Function to give description of autocorrelation result
}
\details{
Call via \code{AutoGradeDescription(x)}
}
\author{
Revanth Nemani, \email{raynemani@gmail.com}
}
